Fix bugs, add world.c contents & first test for hashmap

+++ libft/hash_map.c
@@ -6,12 +6,13 @@
 /*   By: ohakola <ohakola@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/12 15:53:35 by ohakola           #+#    #+#             */
-/*   Updated: 2020/09/12 17:35:15 by ohakola          ###   ########.fr       */
+/*   Updated: 2020/09/15 19:57:51 by ohakola          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "hash_map.h"
 #include "hash_map_utils.h"
+#include "ft_printf.h"
 
 /*
 ** Creates a hash table of given size.
@@ -21,7 +22,7 @@
 t_hash_table		*hash_map_create(int size)
 {
 	t_hash_table	*table;
-	size_t			i;
+	int				i;
 
 	if (!(table = malloc(sizeof(t_hash_table))))
 	{
