+++ libft/hash_map_delete_free.c
@@ -6,7 +6,7 @@
 /*   By: ohakola <ohakola@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/12 17:25:46 by ohakola           #+#    #+#             */
-/*   Updated: 2020/09/12 17:35:08 by ohakola          ###   ########.fr       */
+/*   Updated: 2020/09/15 19:55:22 by ohakola          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -42,7 +42,7 @@ void				hash_map_delete_free(t_hash_table *table, int key)
 
 void				hash_map_clear_free(t_hash_table *table)
 {
-	size_t		i;
+	int			i;
 	t_hash_node	*temp;
 	t_hash_node	*curr;
